requirements.txt
ffmpeg-python
tqdm
pandas
yt-dlp
pydub
python-dotenv
torchaudio
gradio
pytube
seaborn